Some minor refactoring. CameraPoseEstimator::Update can now take a cv::Mat, some dead code was removed, some signed/unsigned comparison warnings resolved.

+++ src/Obstacle.hpp
@@ -39,7 +39,7 @@ struct Obstacle
     for (auto coord : _coords)
     {
       result += "(";
-      for (int i = 0; i < coord.size(); i++)
+      for (size_t i = 0; i < coord.size(); i++)
       {
         result += std::to_string(coord[i]);
         if (i < coord.size() - 1)
